README.md
An analysis focused on the qualitative variables of tweet content and sentiment. But to complement this evaluation, the quantitative fluctuation of likes and retweets was also considered to show a positive correlation between content and popularity. Therefore, the data collection method primarily involved observation, specifically targeting Elon Musk’s tweets during the September 2023 timeframe.